feat(api): 多用户数据隔离与归属校验全链
- store: List/Create/Get 按 userID 过滤（DB + 内存双路径），
  telegram 配置 per-user map 化，decisions/orders/fills/positions/
  equities 聚合端点按 trader 属主过滤
- service: 全部资源方法透传 userID，Get/Update/Delete/Activate/
  Duplicate/ClosePosition 归属校验（空 UserID 历史数据放行）
- handler: currentUserID 取 JWT Subject；数据端点 authorizeTrader；
  回测/辩论引擎方法归属校验（OwnsRun + Get 前置）；batch/stream/
  klines/fills 补防越权
- engine: CredentialsResolver.Resolve 按属主用户解析账户；
  loop 退出路径 defer 清理 runner
- dto: PositionDTO 全字段（修复数量/标记价/浮盈裁剪）、
  EquityPointDTO 语义修正、hyperliquid 私钥加密落库字段
- test: 隔离断言 iso_test（DB 双路径）

+++ internal/trader/engine/ordersync.go
@@ -92,9 +92,10 @@ func (e *Engine) orderSync(ctx context.Context, t *model.Trader) error {
 // 简化实现：开仓成交补录 OPEN 持仓（本地已存在则跳过）；平仓成交结算本地持仓。
 func (e *Engine) positionBuilder(traderID string, f exchange.Fill) {
 	open := e.store.ListPositionsByTrader(traderID)
-	side := "long"
+	// 成交方向的反向 = 待结算持仓方向：sell 平多 / buy 平空
+	side := "short"
 	if f.Side == "sell" {
-		side = "short"
+		side = "long"
 	}
 	// 平仓侧：find 同 symbol 的 OPEN 持仓（方向匹配）→ 结算
 	var matched *model.Position
